Extract coop TUI viewport layout helpers
Committed-By-Agent: codex
